Fixes Update New version

+++ platform/plugins/ecommerce/src/Http/Controllers/ShippingMethodController.php
@@ -6,12 +6,14 @@ use Assets;
 use Botble\Base\Events\DeletedContentEvent;
 use Botble\Base\Http\Controllers\BaseController;
 use Botble\Base\Http\Responses\BaseHttpResponse;
+use Botble\Ecommerce\Enums\ShippingRuleTypeEnum;
 use Botble\Ecommerce\Http\Requests\AddShippingRegionRequest;
 use Botble\Ecommerce\Http\Requests\ShippingRuleRequest;
 use Botble\Ecommerce\Repositories\Interfaces\OrderInterface;
 use Botble\Ecommerce\Repositories\Interfaces\ShippingInterface;
 use Botble\Ecommerce\Repositories\Interfaces\ShippingRuleInterface;
 use Botble\Ecommerce\Repositories\Interfaces\ShippingRuleItemInterface;
+use EcommerceHelper;
 use Exception;
 use Illuminate\Contracts\View\Factory;
 use Illuminate\Http\Request;
@@ -21,32 +23,14 @@ use Throwable;
 
 class ShippingMethodController extends BaseController
 {
-    /**
-     * @var ShippingInterface
-     */
-    protected $shippingRepository;
+    protected ShippingInterface $shippingRepository;
 
-    /**
-     * @var ShippingRuleInterface
-     */
-    protected $shippingRuleRepository;
+    protected ShippingRuleInterface $shippingRuleRepository;
 
-    /**
-     * @var OrderInterface
-     */
-    protected $orderRepository;
+    protected OrderInterface $orderRepository;
 
-    /**
-     * @var ShippingRuleItemInterface
-     */
-    protected $shippingRuleItemRepository;
+    protected ShippingRuleItemInterface $shippingRuleItemRepository;
 
-    /**
-     * @param ShippingInterface $shippingRepository
-     * @param ShippingRuleInterface $shippingRuleRepository
-     * @param OrderInterface $orderRepository
-     * @param ShippingRuleItemInterface $shippingRuleItemRepository
-     */
     public function __construct(
         ShippingInterface $shippingRepository,
         ShippingRuleInterface $shippingRuleRepository,
@@ -67,11 +51,21 @@ class ShippingMethodController extends BaseController
     {
         page_title()->setTitle(trans('plugins/ecommerce::shipping.shipping_methods'));
 
+        if (EcommerceHelper::loadCountriesStatesCitiesFromPluginLocation()) {
+            Assets::addScriptsDirectly('vendor/core/plugins/location/js/location.js');
+        }
+
         Assets::addStylesDirectly(['vendor/core/plugins/ecommerce/css/ecommerce.css'])
             ->addScriptsDirectly(['vendor/core/plugins/ecommerce/js/shipping.js'])
             ->addScripts(['input-mask']);
 
-        $shipping = $this->shippingRepository->allBy([], ['rules']);
+        $shipping = $this->shippingRepository->allBy([], [
+            'rules',
+            'rules.items' => function ($query) {
+                $query->orderBy('created_at', 'desc')
+                    ->orderBy('id', 'desc');
+            },
+        ]);
 
         return view('plugins/ecommerce::shipping.methods', compact('shipping'));
     }
@@ -90,7 +84,7 @@ class ShippingMethodController extends BaseController
             'country' => $country,
         ]);
 
-        if (!$shipping) {
+        if (! $shipping) {
             return $response
                 ->setError()
                 ->setMessage(trans('plugins/ecommerce::shipping.error_when_adding_new_region'));
@@ -114,7 +108,7 @@ class ShippingMethodController extends BaseController
 
         $this->shippingRuleRepository->createOrUpdate([
             'name' => trans('plugins/ecommerce::shipping.delivery'),
-            'type' => 'base_on_price',
+            'type' => ShippingRuleTypeEnum::BASED_ON_PRICE,
             'price' => $price,
             'from' => $from,
             'to' => $to,
@@ -180,22 +174,30 @@ class ShippingMethodController extends BaseController
 
         $rule->fill($request->input());
 
-        $this->shippingRuleRepository->createOrUpdate($rule);
+        $rule = $this->shippingRuleRepository->createOrUpdate($rule);
 
-        $this->shippingRuleItemRepository->deleteBy(['shipping_rule_id' => $id]);
+        if ($rule->type != ShippingRuleTypeEnum::BASED_ON_ZIPCODE) {
+            $this->shippingRuleItemRepository->deleteBy(['shipping_rule_id' => $id]);
 
-        foreach ($request->input('shipping_rule_items', []) as $key => $item) {
-            if (Arr::get($item, 'is_enabled', 0) == 0 || Arr::get($item, 'adjustment_price', 0) != 0) {
-                $this->shippingRuleItemRepository->createOrUpdate([
-                    'shipping_rule_id' => $id,
-                    'city' => $key,
-                    'adjustment_price' => Arr::get($item, 'adjustment_price', 0),
-                    'is_enabled' => Arr::get($item, 'is_enabled', 0),
-                ]);
+            foreach ($request->input('shipping_rule_items', []) as $key => $item) {
+                if (Arr::get($item, 'is_enabled', 0) == 0 || Arr::get($item, 'adjustment_price', 0) != 0) {
+                    $this->shippingRuleItemRepository->createOrUpdate([
+                        'shipping_rule_id' => $id,
+                        'city' => $key,
+                        'adjustment_price' => Arr::get($item, 'adjustment_price', 0),
+                        'is_enabled' => Arr::get($item, 'is_enabled', 0),
+                    ]);
+                }
             }
         }
 
-        return $response->setMessage(trans('core/base::notices.update_success_message'));
+        $data = view('plugins/ecommerce::shipping.rules.item', compact('rule'))->render();
+
+        return $response->setData([
+                'rule' => $rule->toArray(),
+                'html' => $data,
+            ])
+            ->setMessage(trans('core/base::notices.update_success_message'));
     }
 
     /**
@@ -207,10 +209,13 @@ class ShippingMethodController extends BaseController
     public function postCreateRule(ShippingRuleRequest $request, BaseHttpResponse $response)
     {
         $rule = $this->shippingRuleRepository->createOrUpdate($request->input());
-        $data = view('plugins/ecommerce::shipping.rule-item', compact('rule'))->render();
+        $data = view('plugins/ecommerce::shipping.rules.item', compact('rule'))->render();
 
         return $response
             ->setMessage(trans('core/base::notices.create_success_message'))
-            ->setData($data);
+            ->setData([
+                'rule' => $rule->toArray(),
+                'html' => $data,
+            ]);
     }
 }
